Restructring the code for a common base class for the clients. Changes for the discord webhook support
