--- Jellyfin.Plugin.Newsletters/ScheduledTasks/EmailNewsletterTask.cs
@@ -1,56 +1,0 @@
-#pragma warning disable CS1591
-
-using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using Jellyfin.Plugin.Newsletters.Emails.EMAIL;
-using MediaBrowser.Controller.Library;
-using MediaBrowser.Model.Globalization;
-using MediaBrowser.Model.Tasks;
-
-namespace Jellyfin.Plugin.Newsletters.ScheduledTasks
-{
-    /// <summary>
-    /// Class RefreshMediaLibraryTask.
-    /// </summary>
-    public class EmailNewsletterTask : IScheduledTask
-    {
-        /// <inheritdoc />
-        public string Name => "Email Newsletter";
-
-        /// <inheritdoc />
-        public string Description => "Email Newsletters";
-
-        /// <inheritdoc />
-        public string Category => "Newsletters";
-
-        /// <inheritdoc />
-        public string Key => "EmailNewsletters";
-
-        /// <summary>
-        /// Creates the triggers that define when the task will run.
-        /// </summary>
-        /// <returns>IEnumerable{BaseTaskTrigger}.</returns>
-        public IEnumerable<TaskTriggerInfo> GetDefaultTriggers()
-        {
-            yield return new TaskTriggerInfo
-            {
-                Type = TaskTriggerInfo.TriggerInterval,
-                IntervalTicks = TimeSpan.FromHours(168).Ticks
-            };
-        }
-
-        /// <inheritdoc />
-        public Task ExecuteAsync(IProgress<double> progress, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            progress.Report(0);
-
-            Smtp mySmtp = new Smtp();
-            mySmtp.SendEmail();
-            progress.Report(100);
-            return Task.CompletedTask;
-        }
-    }
-}
